+++ src/main/js/doc.js
@@ -26,7 +26,6 @@
 
 import { reportError, reportFatal, reportWarning } from "./error.js";
 import { ns_ebutts, ns_ittp, ns_itts, ns_tt, ns_ttp, ns_tts } from "./names.js";
-import { createSAXParserFromDOMParser } from "./parser.js";
 import { byName, byQName } from "./styles.js";
 import { ComputedLength, hasOwnProperty, parseLength } from "./utils.js";
 
@@ -82,7 +81,7 @@ import { ComputedLength, hasOwnProperty, parseLength } from "./utils.js";
  * @returns {?TT} Opaque in-memory representation of an IMSC1 document
  */
 
-export function fromXML(xmlstring, errorHandler, metadataHandler, parser = createSAXParserFromDOMParser()) {
+export function fromXML(xmlstring, errorHandler, metadataHandler, parser) {
     const estack = [];
     const xmllangstack = [];
     const xmlspacestack = [];
